refactor: rename MULTI_ERROR_HANMDLERS and its related parts
