+++ projects/core/src/error-handling/index.ts
@@ -9,4 +9,4 @@ export * from './effects-error-handler';
 export * from './error-handling.module';
 // export * from './error-interceptors'; TODO: not used, to be removed in separate ticket
 export * from './http-error-handler';
-export * from './multi-error-handlers';
+export * from './multi-error-handler';
