+++ projects/core/src/error-handling/multi-error-handler/index.ts
@@ -5,5 +5,5 @@
  */
 
 export * from './logging-error-handler';
-export * from './multi-error-handlers';
-export * from './provide-multi-error-handlers';
+export * from './multi-error-handler';
+export * from './provide-multi-error-handler';
